Refactored some classes
